+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- add `create_symlink` function to create a symbolic link
+- add `verify_is_symlink` function to check if a file is a symbolic link
 - add `describe_error` function to get human-readable descriptions out of errors
 
 ## v1.6.1 - 28 March 2024